Separate method from generated code to customized code
