+++ service_project.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"errors"
+	"net/http"
+
+	"github.com/ca-risken/core/proto/project"
+)
+
+func (g *gatewayService) createProjectHandler(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	user, err := getRequestUser(r)
+	if err != nil {
+		writeResponse(ctx, w, http.StatusUnauthorized, map[string]interface{}{errorJSONKey: errors.New("InvalidUser")})
+	}
+	req := &project.CreateProjectRequest{}
+	req.UserId = user.userID // force update by own userID
+	bind(req, r)
+	if err := req.Validate(); err != nil {
+		writeResponse(ctx, w, http.StatusBadRequest, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	resp, err := g.projectClient.CreateProject(ctx, req)
+	if err != nil {
+		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	writeResponse(ctx, w, http.StatusOK, map[string]interface{}{successJSONKey: resp})
+}
