+++ service_project_generated.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"errors"
 	"net/http"
 
 	"github.com/ca-risken/core/proto/project"
@@ -16,27 +15,6 @@ func (g *gatewayService) listProjectHandler(w http.ResponseWriter, r *http.Reque
 		return
 	}
 	resp, err := g.projectClient.ListProject(ctx, req)
-	if err != nil {
-		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
-		return
-	}
-	writeResponse(ctx, w, http.StatusOK, map[string]interface{}{successJSONKey: resp})
-}
-
-func (g *gatewayService) createProjectHandler(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	user, err := getRequestUser(r)
-	if err != nil {
-		writeResponse(ctx, w, http.StatusUnauthorized, map[string]interface{}{errorJSONKey: errors.New("InvalidUser")})
-	}
-	req := &project.CreateProjectRequest{}
-	req.UserId = user.userID // force update by own userID
-	bind(req, r)
-	if err := req.Validate(); err != nil {
-		writeResponse(ctx, w, http.StatusBadRequest, map[string]interface{}{errorJSONKey: err.Error()})
-		return
-	}
-	resp, err := g.projectClient.CreateProject(ctx, req)
 	if err != nil {
 		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
 		return
